lambda/libs/ocr_tools/requirements.txt
amazon_textract_prettyprinter==0.0.13
boto3==1.24.4
botocore==1.27.17
charset_normalizer==2.0.12
fastapi==0.78.0
mangum==0.15.0
Pillow==9.1.1
pydantic==1.9.1
pytest==7.1.2
python_magic==0.4.27
setuptools==56.0.0
uvicorn==0.18.1
